DBZ-234 Move exceptions out of separate package
