+++ debezium-api/src/main/java/io/debezium/DebeziumException.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.errors;
+package io.debezium;
 
 /**
  * Base exception returned by Debezium API.
