+++ debezium-api/src/main/java/io/debezium/engine/StopEngineException.java
@@ -3,7 +3,9 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.errors;
+package io.debezium.engine;
+
+import io.debezium.DebeziumException;
 
 /**
  * An exception that is used to tell the engine to process the last source record and to then stop. When raised by
